requirements.txt
pyyaml>=4.2b1
selenium==3.13.0
argparse==1.4.0